Make code closer to original dcraw.c code. These type casts are not necessary for avoiding compiler warnings with newer versions of GCC.

+++ dcraw_api.cc
@@ -166,15 +166,15 @@ int dcraw_image_dimensions(dcraw_data *raw, int flip, int *height, int *width)
 {
     if (raw->fuji_width) {
 	/* Copied from DCRaw's fuji_rotate() */
-	*width = (int)((raw->fuji_width - 1) / raw->fuji_step);
-	*height = (int)((raw->height - raw->fuji_width + 1) / raw->fuji_step);
+	*width = (raw->fuji_width - 1) / raw->fuji_step;
+	*height = (raw->height - raw->fuji_width + 1) / raw->fuji_step;
     } else {
 	if (raw->pixel_aspect < 1)
-	    *height = (int)(raw->height / raw->pixel_aspect + 0.5);
+	    *height = raw->height / raw->pixel_aspect + 0.5;
 	else
 	    *height = raw->height;
 	if (raw->pixel_aspect > 1)
-	    *width = (int)(raw->width * raw->pixel_aspect + 0.5);
+	    *width = raw->width * raw->pixel_aspect + 0.5;
 	else
 	    *width = raw->width;
     }
